stage 2: open_table: create local temporary table from global for DML + SHOW
On open, a local temporary table copy is created from a global TABLE_SHARE.
It holds a reference to that share, and holds the increased tdc reference count.

Also, it holds an explicit MDL lock to the duration of local copy's existence.
The latter becomes important for ALTER and DROP.

TRUNCATE drops a local copy.

When the local copy is deleted (by TRUNCATE or SHOW), the reference is
decreased. The SHOW part requires some clarification, but I believe that we can
implement dropping a local table as soon as its record count reaches 0 at the
end of transaction.

SHOW CREATE now also uses a local temporary table for showing to simplify the
implementation (otherwise, it'd need to support displaying both global and
local descriptors, which differ in minor things)