MENT-2235 Aria engine: log initialization failed
Some thing causes the aria_log_control file to be larger than the expected
52 bytes. The control file has the correct information but somehow it
is filled up with ox00 bytes up to 512 bytes.
This could have happened in case of a file system crash that enlarged
the file to the sector boundary.

Fixed that aria will ignore bytes outside of it's expected

Other things:
- Fixed wrong DBUG_ASSERT() in my_malloc_size_cb_func() that could cause
  crashes in debug binaries during Aria recovery.